Fixed tests, moved the relevant GroundControl tests to the EditorPublishButton component
