+++ client/post-editor/editor-ground-control/test/index.jsx
@@ -2,10 +2,8 @@
  * External dependencies
  */
 import { expect } from 'chai';
-import moment from 'moment';
 import { noop } from 'lodash';
 import React from 'react';
-import sinon from 'sinon';
 import mockery from 'mockery';
 
 /**
@@ -23,16 +21,6 @@ const MOCK_SITE = {
 		publish_posts: true
 	},
 	options: {}
-};
-
-const MOCK_USER = {
-	email_verified: true
-};
-
-const MOCK_USER_UTILS = {
-	needsVerificationForSite: function( site ) { // eslint-disable-line no-unused-vars
-		return ! MOCK_USER.email_verified;
-	}
 };
 
 describe( 'EditorGroundControl', function() {
@@ -57,9 +45,6 @@ describe( 'EditorGroundControl', function() {
 		mockery.registerMock( 'lib/posts/stats', {
 			recordEvent: noop,
 			recordStat: noop
-		} );
-		mockery.registerMock( 'lib/user/utils', {
-			needsVerificationForSite: () => ! MOCK_USER.email_verified,
 		} );
 		EditorGroundControl = require( '../' );
 
@@ -94,154 +79,6 @@ describe( 'EditorGroundControl', function() {
 			).instance();
 
 			expect( tree.getPreviewLabel() ).to.equal( 'Preview' );
-		} );
-	} );
-
-	describe( '#getPrimaryButtonLabel()', function() {
-		it( 'should return Update if the post was originally published and is still slated to be published', function() {
-			var tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'publish' } }
-					post={ { status: 'publish' } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Update' );
-		} );
-
-		it( 'should return Update if the post was originally published and is currently reverted to non-published status', function() {
-			var tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'publish' } }
-					post={ { status: 'draft' } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Update' );
-		} );
-
-		it( 'should return Schedule if the post is dated in the future and not scheduled', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft' } }
-					post={ { date: nextMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Schedule' );
-		} );
-
-		it( 'should return Schedule if the post is dated in the future and published', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft' } }
-					post={ { date: nextMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Schedule' );
-		} );
-
-		it( 'should return Update if the post is scheduled and dated in the future', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'future', date: nextMonth } }
-					post={ { title: 'change', status: 'future', date: nextMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Update' );
-		} );
-
-		it( 'should return Update if the post is scheduled, dated in the future, and next status is draft', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'future', date: nextMonth } }
-					post={ { title: 'change', status: 'draft', date: nextMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Update' );
-		} );
-
-		it( 'should return Publish if the post is scheduled and dated in the past', function() {
-			var now = moment( new Date() ),
-				lastMonth = now.month( now.month() - 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'future', date: lastMonth } }
-					post={ { title: 'change', status: 'future', date: lastMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Publish' );
-		} );
-
-		it( 'should return Update if the post was originally published and is scheduled with date in the past', function() {
-			var now = moment(),
-				lastMonth = now.month( now.month() - 1 ).format(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'publish', date: lastMonth } }
-					post={ { title: 'change', status: 'future', date: lastMonth } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Update' );
-		} );
-
-		it( 'should return Publish if the post is a draft', function() {
-			var tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft' } }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Publish' );
-		} );
-
-		it( 'should return "Submit for Review" if the post is a draft and user can\'t publish', function() {
-			var tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft' } }
-					site={ {
-						capabilities: {
-							publish_posts: false
-						}
-					} }
-				/>
-			).instance();
-
-			expect( tree.getPrimaryButtonLabel() ).to.equal( 'Submit for Review' );
 		} );
 	} );
 
@@ -312,190 +149,6 @@ describe( 'EditorGroundControl', function() {
 			var tree = shallow( <EditorGroundControl post={ {} } hasContent={ false } /> ).instance();
 
 			expect( tree.isPreviewEnabled() ).to.be.false;
-		} );
-	} );
-
-	describe( '#isPrimaryButtonEnabled()', function() {
-		it( 'should return true if form is not publishing and post is not empty', function() {
-			var tree = shallow( <EditorGroundControl isPublishing={ false } post={ {} } hasContent isDirty isNew /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.true;
-		} );
-
-		it( 'should return false if form is not publishing and post is not empty, but user is not verified', function() {
-			MOCK_USER.email_verified = false;
-			let tree = shallow( <EditorGroundControl isPublishing={ false } post={ {} } user={ MOCK_USER } userUtils={ MOCK_USER_UTILS } hasContent isDirty isNew /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.false;
-			MOCK_USER.email_verified = true;
-		} );
-
-		it( 'should return true if form is not publishind and post is new and has content, but is not dirty', function() {
-			var tree = shallow( <EditorGroundControl isPublishing={ false } post={ {} } hasContent isDirty={ false } isNew /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.true;
-		} );
-
-		it( 'should return false if form is publishing', function() {
-			var tree = shallow( <EditorGroundControl isPublishing /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.false;
-		} );
-
-		it( 'should return false if saving is blocked', function() {
-			var tree = shallow( <EditorGroundControl isSaveBlocked /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.false;
-		} );
-
-		it( 'should return false if not dirty and has no content', function() {
-			var tree = shallow( <EditorGroundControl post={ {} } isDirty={ false } hasContent={ false } isNew /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.false;
-		} );
-
-		it( 'should return false if post has no content', function() {
-			var tree = shallow( <EditorGroundControl post={ {} } hasContent={ false } /> ).instance();
-
-			expect( tree.isPrimaryButtonEnabled() ).to.be.false;
-		} );
-	} );
-
-	describe( '#onPrimaryButtonClick', function() {
-		it( 'should publish a draft', function() {
-			var onPublish = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					post={ { status: 'draft' } }
-					site={ MOCK_SITE }
-					onPublish={ onPublish }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onPublish ).to.have.been.called;
-		} );
-
-		it( 'should schedule a posted dated in future', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				onPublish = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft', date: nextMonth } }
-					post={ { title: 'change', status: 'draft', date: nextMonth } }
-					onPublish={ onPublish }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onPublish ).to.have.been.called;
-		} );
-
-		it( 'should save a scheduled post dated in future', function() {
-			var now = moment( new Date() ),
-				nextMonth = now.month( now.month() + 1 ).format(),
-				onPublish = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'future', date: nextMonth } }
-					post={ { title: 'change', status: 'future', date: nextMonth } }
-					onPublish={ onPublish }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onPublish ).to.have.been.called;
-		} );
-
-		it( 'should publish a scheduled post dated in past', function() {
-			var now = moment( new Date() ),
-				lastMonth = now.month( now.month() - 1 ).format(),
-				onPublish = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'future', date: lastMonth } }
-					post={ { title: 'change', status: 'future', date: lastMonth } }
-					onPublish={ onPublish }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onPublish ).to.have.been.called;
-		} );
-
-		it( 'should update a published post that has changed status', function() {
-			var onSave = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'publish' } }
-					post={ { title: 'change', status: 'draft' } }
-					onSave={ onSave }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onSave ).to.have.been.called;
-		} );
-
-		it( 'should update a published post scheduled in the past', function() {
-			var now = moment( new Date() ),
-				lastMonth = now.month( now.month() - 1 ).format(),
-				onSave = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'publish' } }
-					post={ { title: 'change', status: 'future', date: lastMonth } }
-					onSave={ onSave }
-					site={ MOCK_SITE }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onSave ).to.have.been.called;
-		} );
-
-		it( 'should set status to "pending" if the user can\'t publish', function() {
-			var onSave = sinon.spy(),
-				tree;
-
-			tree = shallow(
-				<EditorGroundControl
-					savedPost={ { status: 'draft' } }
-					onSave={ onSave }
-					site={ {
-						capabilities: {
-							publish_posts: false
-						}
-					} }
-				/>
-			).instance();
-
-			tree.onPrimaryButtonClick();
-
-			expect( onSave ).to.have.been.calledWith( 'pending' );
 		} );
 	} );
 } );
